+++ frontend-react/e2e/helpers/utils.ts
@@ -1,6 +1,5 @@
 import { expect, Page } from "@playwright/test";
 import fs from "node:fs";
-import { fromDateWithTime, toDateWithTime } from "../pages/daily-data";
 
 export const TEST_ORG_IGNORE = "ignore";
 export const TEST_ORG_UP_RECEIVER_FULL_ELR = "FULL_ELR";
@@ -102,4 +101,38 @@ export async function tableColumnDateTimeInRange(
         }
     }
     return datesInRange;
+}
+
+export function fromDateWithTime(date: string, time: string) {
+    const fromDateTime = new Date(date);
+
+    if (time) {
+        // eslint-disable-next-line prefer-const
+        let [hours, minutes] = time
+            .substring(0, time.length - 2)
+            .split(":")
+            .map(Number);
+        hours = hours + (time.indexOf("pm") !== -1 ? 12 : 0);
+        fromDateTime.setHours(hours, minutes, 0, 0);
+    } else {
+        fromDateTime.setHours(0, 0, 0);
+    }
+    return fromDateTime;
+}
+
+export function toDateWithTime(date: string, time: string) {
+    const toDateTime = new Date(date);
+
+    if (time) {
+        // eslint-disable-next-line prefer-const
+        let [hours, minutes] = time
+            .substring(0, time.length - 2)
+            .split(":")
+            .map(Number);
+        hours = hours + (time.indexOf("pm") !== -1 ? 12 : 0);
+        toDateTime.setHours(hours, minutes, 0, 0);
+    } else {
+        toDateTime.setHours(23, 59, 0);
+    }
+    return toDateTime;
 }
